chore: move notification processor filters to common package
