+++ .changeset/weak-emus-doubt.md
@@ -1,0 +1,7 @@
+---
+'@backstage/plugin-notifications-backend-module-email': patch
+'@backstage/plugin-notifications-common': patch
+'@backstage/plugin-notifications-node': patch
+---
+
+Move notification processor filter parsing to common package
