+++ plugins/notifications-common/api-report.md
@@ -3,6 +3,13 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
+import { Config } from '@backstage/config';
+
+// @public (undocumented)
+export const getProcessorFiltersFromConfig: (
+  config: Config,
+) => NotificationProcessorFilters;
+
 // @public (undocumented)
 export type NewNotificationSignal = {
   action: 'new_notification';
@@ -31,6 +38,13 @@ export type NotificationPayload = {
   topic?: string;
   scope?: string;
   icon?: string;
+};
+
+// @public (undocumented)
+export type NotificationProcessorFilters = {
+  minSeverity?: NotificationSeverity;
+  maxSeverity?: NotificationSeverity;
+  excludedTopics?: string[];
 };
 
 // @public (undocumented)
