+++ plugins/notifications-common/package.json
@@ -32,6 +32,7 @@
     "test": "backstage-cli package test"
   },
   "dependencies": {
+    "@backstage/config": "workspace:^",
     "@material-ui/icons": "^4.9.1"
   },
   "devDependencies": {
