+++ plugins/notifications-common/src/index.ts
@@ -22,3 +22,4 @@
 
 export * from './types';
 export * from './constants';
+export * from './filters';
